Adding a Game Dedicated Project
I've added a project dedicated to the game, but it doesn't import the files properly. Worst case scenario we have to copy things over, but I'm just leaving it empty for now. Oh, and I also changed the name of the engine folder to 'pc' since that's the namespace we're using.