Code Cleanup
Cleaned up some of the code by changing function names so that they're clearer. I also set up a vector to get meshes from the objects directly since the mesh vector we have makes a copy of all that data. Collision still needs to be setup, and there's still other cleaning up to do, but that's probably all I'm doing for today.