Collisions Working
Collisions have been added, though because of the weird way our positions and whatnot work, it's not clear where the hitboxes are. The project really needs to be cleaned up and organized better, which I hope to do either Friday or Saturday. If you want to see when collisions happen, uncomment the word "Hit!" in the update section so that it prints when collision happens.

NOTE: make sure you change your C++ version to c++17 in order for the project to work. There's a screenshot in the Discord #code-general for how to do it. If you can't find it, it's just 'Debug > GDW_Y2 - BF PJT Properties... > Configuration Manager > C++ Language Standard', and set to "ISO C++ 17".