bomb ass collision
its basically perfect. except for one detail. if you collide on the right side of an object and move up while still holding down D, you wont be able to go right even after clearing the bounding box of the obstacle. this happens in every direction, not just right. Not sure what the cause is but I can look into it later. For now this is more than adequate for our game.