Message About Recent Changes
 So basically although the player object was being made in the map, there was nothing bringing it to the actual scene. Or if there was, that somehow broke when I changed the model the player uses (which had to be done because our 1st year model is different from our second year model). I fixed this by adding in a function for getting the player object from the map.
As for the player's physics body, that had to be added in manually because that got lost somehow when making the new player object. Being honest this should all just be a function in the player class instead, but that's a change for another time.
As for loading in the animations, that had to be done manually. Honestly it'd probably be best if we had directions to the animations in a text file and had it read that instead, but once again that's another matter for another time.
I have a few things I want us to try and fix or improve for GameCon, but I'll prepare that list once the Computer Graphics assignment is done, and maybe the Algorithms midterm.