Loading Screen (WIP)
Creating a loading screen ,though it's a work and progress and just shows up black. I also added a button for going to the ranking system (it's the biggest button near the bottom of the screen), but it crashes sometimes when going back to the main menu, so I need to fix that. The text positioning system is also broken, and the text is upside down, so I need to fix that too.