Updating utils folder
The items in the utils folder have been updated with the most recent content. I'll get to working on the game code itself, but there's still some stuff I want to make for the utils project. Though I may put that stuff on the back burner and focus on just the game.