Adding in optimized animations.
The animations have been optimized for the game now. Not all frames have been added in, but the animations that were already setup are all in. I might wait a bit to keep working on the game though to see if any news about the GameCon comes out this week.