Working on Fixing Shader Error
I may have found what was causing the shaders to break. Basically data is getting held over, but it keeps getting rewritten. So I'm going to try and fix that. Furthermore, I've been trying to replace the arrow with an image since that's less intensive. Finally, I added another object file that has the objects merged.