Shader Problem Fixed Via Workaround
The scene switch problem seems to have been fixed via a work around. Basically I just reset the scene everytime and rebuild it from scratch instead of opening and closing it over and over again. I wish I thought of this sooner, but hindsight is 2020. I don't really have anything else to do, so while I have about a month and a half left to help fix this thing, I'll try to improve it as much as I can.