Score Entry Works Now
The score entry for the game works now. The formatting needs work, but it's all good for now. I just need to make the game over go to the game over screen rather than the main menu, and the main mechanics will all be in the game in some capacity.