Tweaking Settings, and deleting Blend Files
I was working on tweaking different settings to try and improve the game. And I deleted the .blend files in the project since those don't do anything . There are probably obj files that we still don't use, but I'll probably leave all that in.

I need to adjust the imGui window since it's too big, and sometimes doesn't disappear like it's supposed to.