Improving Terrain
Terrain has been improved to allow for weights to mix textures together rather than the weights being fixed. I'm going to use this for background objects in the menu. The city texture is rough for now, but if I have time I'll go back and add in more. I'll probably just keep to using terrain since the more I add will make the game run worse. I also want to change the skyboxes and improve upon them. I'll probably make something that stays under the player the whole time to fill in the empty space. I also need to figure out how to shift the platforms for the player.