Moving Props into Proper Folder
All the prop objects have been moved to the right folder, and all the file paths should be correct. I also fixed some objects that didn't have textures and/or had their origins in the wrong place. As for projectiles, the hitboxes need to be boosted to allow for hits to happen. The crash that sometimes happens when an enemy is killed still needs to be fixed.