Debug Map Added
A debug map has been added, which uses all items. I think the line amount of incorrect, but I'll fix that later. I've also improved the formatting for the ranking scene. I'm probably going to use terrain and water in that scene given that I don't have water anyway, and it's the only object type we don't use. I probably added a lot more, but I honestly don't remember what it all is.