Improved Dash
The dash code has been optimized and improved so that the charge up time can be adjusted. I also replaced the 3D arrow with a flat 2D arrow so that less polygons are used. I'm going to adjust the time slow effect in a similar fashion so that can be more easily adjusted. It also should have a cool down of some kind.