Final Work for the Summer (Removing Unused Code)
Final work for the summer on the game. Maps 2 and 3 don't load completely, so that may be a row/col limiter problem or an incorrect numbering problem. Either way, I'm going to work on deleting unnecessary comments, updating them, and getting rid of commented out code that isn't used anymore. So revert back to here if necessary.